requirements.txt
llama-cpp-python==0.3.20
httpx==0.27.0
PyNaCl==1.5.0
beautifulsoup4==4.12.3
trafilatura==1.12.0
pyyaml==6.0.2
cryptography>=41.0.0
